ADC Review 1
1. unrouted trace 1: Test Point 3VCCA is floating. It may be corrected
by providing a multi-layer via through the pad so it hits Internal Plane
2 (split plane 3VCCA).
2. unrouted trace 2: Pin CCW at 2220, 200 is floating. It must be
connected to 3VCCA.
3. Add trace from the AGND via below the power connector to gnd pin 1 of
the connector, for the single-point ground connection. Watch out for
unintended effects when repouring copper.
4. Internal plane 2 bottom left, change designation '3VCCA' to 5V0.
Might as well move '3VCCA' to split plane around BGA.
5. Re-enable Top Overlay before gerber file creation.
6. Many overlay designations need to be moved out of the way of vias and
pads.
7. To obtain a 100 ohm impedance, each differential (+ and -) trace
needs to be at a 50 ohm impedance. Using the impedance calculator in the
board layer stack manager menu, the required width is 9.56 mils
(rounded) using the Bittele 8 layer stackup table for the outer layers
(microstrip).

For the inner layers (asymmetric stripline), the required width is:
14 mils for mid-layer 1
15.4 mils for mid-layer 2
14.3 mils for mid-layer 3
14 mils for mid-layer 4
Calculators:
http://www.hughescircuits.com/index.php/en/support/impedance-calculator
8. Length adjustments for slower propogation delays within a dielectric
have not yet been made. Do they need to?
9. Board looks great. Nice improvements.
10. Will continue to check. Please send new commits for ADC and data
formatter as soon as possible!